mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.67.0",
    "docx>=0.2.4",
    "fitz>=0.0.1.dev2",
    "flask>=3.1.2",
    "mcp>=1.14.0",
    "pymupdf>=1.26.4",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
]
